Example Axiom Report Data

A list containing example Axiom genotyping report data with header information and quality metrics from the Axiom genotyping platform.

Usage

example_axiom_report

Format

A list with 2 elements:

header

Character vector with 9 elements containing metadata including chip type, library set information, processing parameters, and timestamps

data

Data frame with 15 rows and 4 variables:

cel_files: Character. CEL file names

total_call_rate: Numeric. Percentage of successful genotype calls

average_heterozygosity: Numeric. Average heterozygosity rate

median_intensity: Numeric. Median signal intensity

Examples

data(example_axiom_report)
example_axiom_report$header
#> [1] "#%chip_type=Axiom_GW_Hu_SNP"                                   
#> [2] "#%lib_set_name=Axiom_GW_Hu_SNP.r2"                             
#> [3] "#%lib_set_version=r2"                                          
#> [4] "#%affymetrix-algorithm-param-apt-time-str=May 23 14:30:45 2025"
#> [5] "#%create_date=2024-01-15T10:30:45Z"                            
#> [6] "#%guid=1234567890abcdef"                                       
#> [7] "#%program_name=apt-probeset-genotype"                          
#> [8] "#%program_version=1.21.0"                                      
#> [9] "#%program_company=Affymetrix"                                  
head(example_axiom_report$data)
#>        cel_files total_call_rate average_heterozygosity median_intensity
#> 1 Sample_001.CEL            97.5                  0.230             2150
#> 2 Sample_002.CEL            94.2                  0.285             1890
#> 3 Sample_003.CEL            98.1                  0.215             2340
#> 4 Sample_004.CEL            96.8                  0.250             2100
#> 5 Sample_005.CEL            93.5                  0.305             1750
#> 6 Sample_006.CEL            99.2                  0.220             2500
summary(example_axiom_report$data$total_call_rate)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   92.10   94.50   96.30   96.05   97.70   99.20